requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
crewai>=1.9.0
groq>=0.4.0
notion-client>=2.2.0
apscheduler>=3.10.0
pyyaml>=6.0
dnspython>=2.4.0
email-validator>=2.1.0
python-dotenv>=1.0.0
lxml>=5.0.0
